+++ android/src/main/kotlin/be/tramckrijte/workmanager/WorkmanagerCallHandler.kt
@@ -6,6 +6,7 @@ import androidx.work.Data
 import androidx.work.ExistingPeriodicWorkPolicy
 import androidx.work.ExistingWorkPolicy
 import androidx.work.OneTimeWorkRequest
+import androidx.work.OutOfQuotaPolicy
 import androidx.work.PeriodicWorkRequest
 import androidx.work.WorkManager
 import be.tramckrijte.workmanager.BackgroundWorker.Companion.DART_TASK_KEY
@@ -113,6 +114,7 @@ private object RegisterTaskHandler : CallHandler<WorkManagerCall.RegisterTask> {
             initialDelaySeconds = convertedCall.initialDelaySeconds,
             constraintsConfig = convertedCall.constraintsConfig,
             backoffPolicyConfig = convertedCall.backoffPolicyConfig,
+            outOfQuotaPolicy = convertedCall.outOfQuotaPolicy,
             payload = convertedCall.payload
         )
     }
@@ -131,6 +133,7 @@ private object RegisterTaskHandler : CallHandler<WorkManagerCall.RegisterTask> {
             initialDelaySeconds = convertedCall.initialDelaySeconds,
             constraintsConfig = convertedCall.constraintsConfig,
             backoffPolicyConfig = convertedCall.backoffPolicyConfig,
+            outOfQuotaPolicy = convertedCall.outOfQuotaPolicy,
             payload = convertedCall.payload
         )
     }
@@ -185,6 +188,7 @@ object WM {
         existingWorkPolicy: ExistingWorkPolicy = defaultOneOffExistingWorkPolicy,
         initialDelaySeconds: Long = defaultInitialDelaySeconds,
         constraintsConfig: Constraints = defaultConstraints,
+        outOfQuotaPolicy: OutOfQuotaPolicy? = defaultOutOfQuotaPolicy,
         backoffPolicyConfig: BackoffPolicyTaskConfig?
     ) {
         val oneOffTaskRequest = OneTimeWorkRequest.Builder(BackgroundWorker::class.java)
@@ -200,7 +204,10 @@ object WM {
                     )
                 }
             }
-            .apply { tag?.let(::addTag) }
+            .apply {
+                tag?.let(::addTag)
+                outOfQuotaPolicy?.let(::setExpedited)
+            }
             .build()
         context.workManager()
             .enqueueUniqueWork(uniqueName, existingWorkPolicy, oneOffTaskRequest)
@@ -217,6 +224,7 @@ object WM {
         existingWorkPolicy: ExistingPeriodicWorkPolicy = defaultPeriodExistingWorkPolicy,
         initialDelaySeconds: Long = defaultInitialDelaySeconds,
         constraintsConfig: Constraints = defaultConstraints,
+        outOfQuotaPolicy: OutOfQuotaPolicy? = defaultOutOfQuotaPolicy,
         backoffPolicyConfig: BackoffPolicyTaskConfig?
     ) {
         val periodicTaskRequest =
@@ -237,7 +245,10 @@ object WM {
                         )
                     }
                 }
-                .apply { tag?.let(::addTag) }
+                .apply {
+                    tag?.let(::addTag)
+                    outOfQuotaPolicy?.let(::setExpedited)
+                }
                 .build()
         context.workManager()
             .enqueueUniquePeriodicWork(uniqueName, existingWorkPolicy, periodicTaskRequest)