+++ test/json_mapper_helper_test.dart
@@ -58,6 +58,7 @@ void main() {
             'requiresStorageNotLow': null,
             'backoffPolicyType': null,
             'backoffDelayInMilliseconds': null,
+            'outOfQuotaPolicy': null,
             'inputData': null
           },
         );
@@ -87,6 +88,7 @@ void main() {
             'requiresStorageNotLow': null,
             'backoffPolicyType': null,
             'backoffDelayInMilliseconds': null,
+            'outOfQuotaPolicy': null,
             'inputData': null
           },
         );
@@ -118,6 +120,7 @@ void main() {
               'requiresStorageNotLow': null,
               'backoffPolicyType': null,
               'backoffDelayInMilliseconds': 1000,
+              'outOfQuotaPolicy': null,
               'inputData': null,
             });
       });
@@ -158,6 +161,7 @@ void main() {
               'requiresStorageNotLow': false,
               'backoffPolicyType': 'linear',
               'backoffDelayInMilliseconds': 3000,
+              'outOfQuotaPolicy': null,
               'inputData': jsonEncode({"key": "value"}),
             });
       });
@@ -188,6 +192,7 @@ void main() {
             'requiresStorageNotLow': true,
             'backoffPolicyType': null,
             'backoffDelayInMilliseconds': 1000,
+            'outOfQuotaPolicy': null,
             'inputData': null,
           }
         ],
@@ -214,6 +219,7 @@ void main() {
             'requiresStorageNotLow': false,
             'backoffPolicyType': null,
             'backoffDelayInMilliseconds': 1000,
+            'outOfQuotaPolicy': null,
             'inputData': null,
           }
         ],
@@ -234,6 +240,7 @@ void main() {
             'requiresStorageNotLow': null,
             'backoffPolicyType': null,
             'backoffDelayInMilliseconds': 1000,
+            'outOfQuotaPolicy': null,
             'inputData': null,
           }
         ],
@@ -260,6 +267,7 @@ void main() {
             'requiresStorageNotLow': true,
             'backoffPolicyType': null,
             'backoffDelayInMilliseconds': 1000,
+            'outOfQuotaPolicy': null,
             'inputData': null,
           }
         ],
@@ -286,6 +294,7 @@ void main() {
             'requiresStorageNotLow': true,
             'backoffPolicyType': null,
             'backoffDelayInMilliseconds': 1000,
+            'outOfQuotaPolicy': null,
             'inputData': null,
           }
         ],
@@ -337,6 +346,7 @@ void main() {
               'requiresStorageNotLow': null,
               'backoffPolicyType': null,
               'backoffDelayInMilliseconds': 1000,
+              'outOfQuotaPolicy': null,
               'inputData': null,
             },
           );
@@ -372,6 +382,7 @@ void main() {
               'requiresStorageNotLow': null,
               'backoffPolicyType': backOffPolicy[1],
               'backoffDelayInMilliseconds': 1000,
+              'outOfQuotaPolicy': null,
               'inputData': null,
             },
           );